moving logic to user id module rule
