+++ test/spec/modules/userId_spec.js
@@ -39,6 +39,7 @@ import {extractEids} from '../../../modules/prebidServerBidAdapter/bidderConfig.
 import {generateSubmoduleContainers} from '../../../modules/userId/index.js';
 import { registerActivityControl } from '../../../src/activities/rules.js';
 import { addIdData } from '../../../modules/userId/index.js';
+import { STORAGE_TYPE_COOKIES, STORAGE_TYPE_LOCALSTORAGE, getStorageManager } from '../../../src/storageManager.js';
 
 let assert = require('chai').assert;
 let expect = require('chai').expect;
@@ -3288,4 +3289,51 @@ describe('User ID', function () {
       expect(submoduleNames).to.deep.eql(['modified', 'new']);
     });
   });
+  describe('user id modules - enforceStorageType', () => {
+    let warnLogSpy;
+    const UID_MODULE_NAME = 'userIdModule';
+    const userSync = {
+      userIds: [
+        {
+          name: UID_MODULE_NAME,
+          storage: {
+            type: STORAGE_TYPE_LOCALSTORAGE,
+            name: 'storageName'
+          }
+        }
+      ]
+    };
+
+    beforeEach(() => {
+      warnLogSpy = sinon.spy(utils, 'logWarn');
+    });
+
+    afterEach(() => {
+      warnLogSpy.restore();
+      document.cookie = ''
+    });
+
+    it('should not allow userId module to store data for enforceStorageType set to true', () => {
+      const initialCookie = document.cookie;
+      config.setConfig({
+        userSync: {
+          enforceStorageType: true,
+          ...userSync,
+        }
+      })
+      const storage = getStorageManager({moduleType: MODULE_TYPE_UID, moduleName: UID_MODULE_NAME});
+      storage.setCookie('data', 'value', 20000);
+      expect(initialCookie).to.deep.eql(document.cookie);
+    });
+
+    it('should warn and allow userId module to store data for enforceStorageType unset', () => {
+      const initialCookie = document.cookie;
+      config.setConfig({userSync});
+      const storage = getStorageManager({moduleType: MODULE_TYPE_UID, moduleName: UID_MODULE_NAME});
+      storage.setCookie('cookieName', 'value', 20000);
+      sinon.assert.calledWith(warnLogSpy, `${UID_MODULE_NAME} attempts to store data in ${STORAGE_TYPE_COOKIES} while configuration allows ${STORAGE_TYPE_LOCALSTORAGE}.`);
+      expect(initialCookie).to.not.deep.eql(document.cookie);
+      expect(document.cookie).to.deep.include('cookieName');
+    });
+  })
 });
